same test code for both blockchain implementations
